🔨 Refactor step selection logic and improve preview rendering
