+++ apps/wizard/app_pages/dashboard/app.py
@@ -3,12 +3,13 @@
 import streamlit as st
 from structlog import get_logger
 
-from apps.wizard.app_pages.dashboard.actions import render_action_archive, render_action_execute, render_action_update
 from apps.wizard.app_pages.dashboard.aggrid import make_agrid
-from apps.wizard.app_pages.dashboard.operations import render_operations_list
+from apps.wizard.app_pages.dashboard.operations import render_operations
 from apps.wizard.app_pages.dashboard.preview import render_preview_list
+from apps.wizard.app_pages.dashboard.selection import render_selection_list
 from apps.wizard.app_pages.dashboard.utils import (
     _create_html_button,
+    _get_steps_info,
     check_db,
     load_steps_df,
     load_steps_df_to_display,
@@ -50,12 +51,13 @@ st.title(":material/tv_gen: ETL Dashboard")
 tutorial_html = f"""
 **Example:** Updating a specific Grapher dataset.
 <ol>
-    <li>Select that step from the <i>Steps table</i>.</li>
-    <li>Click on{_create_html_button("Add selected steps to the <i>Operations list</i>", "#002147", "#002147", "#FFFFFF")}.</li>
+    <li>Select the step from the table.</li>
+    <li>The dataset will be added to a preliminary list, where you can preview it.</li>
+    <li>Click on{_create_html_button("Add steps", "#002147", "#002147", "#FFFFFF")}.</li>
     <li>Click on{_create_html_button("Add all dependencies", "#333333", "transparent", "#333333")} (and optionally click on {_create_html_button("Remove non-updateable", "#333333", "transparent", "#333333")}).</li>
-    <li>Click on{_create_html_button("Update X steps", "#002147", "#002147", "white")} to bulk-update them all in one go.</li>
-    <li>Click on{_create_html_button("Replace steps with their latest version", "#333333", "transparent", "#333333")} to populate the <i>Operations list</i> with the newly created steps.</li>
-    <li>Click on{_create_html_button("Run all ETL steps", "#002147", "#002147", "white")} to run the ETL on the new steps.</li>
+    <li>In the <b>Update step</b> section, click on{_create_html_button("Run", "#002147", "#002147", "white")} to bulk-update them all in one go.</li>
+    <li>Click on{_create_html_button("Replace steps with their latest version", "#333333", "transparent", "#333333")} to populate the <b>Selection list</b> with the newly created steps.</li>
+    <li>In the <b>Execute steps</b> section, click on{_create_html_button("Run", "#002147", "#002147", "white")} to run the ETL on the new steps.</li>
     <li>If a step fails, you can manually edit its code and try running ETL again.</li>
 </ol>
 """
@@ -84,6 +86,8 @@ show_all_channels = not st.toggle("Select only grapher and explorer steps", True
 with st.spinner("Loading steps details from ETL and DB..."):
     steps_df = load_steps_df(reload_key=st.session_state["reload_key"])
 
+# Simplify the steps dataframe to show only the relevant columns.
+steps_info = _get_steps_info(steps_df)
 
 ########################################
 # Display STEPS TABLE
@@ -93,28 +97,30 @@ steps_df_display = load_steps_df_to_display(show_all_channels, reload_key=st.ses
 
 # Build and display the grid table with pagination.
 # st.write(steps_df_display.dtypes)
+# with st.container(border=True):
 grid_response = make_agrid(steps_df_display)
 
 ########################################
 # DETAILS LIST
 #
 # Preview of the steps based on user selections.
-# Once happy, the user should click on "Add steps" button and proceed to the "Operations list".
+# Once happy, the user should click on "Add steps" button and proceed to the "Selection list".
 ########################################
 
 
 df_selected = grid_response["selected_rows"]
-render_preview_list(df_selected, steps_df)
+selected_steps = df_selected["step"].tolist() if df_selected is not None else []
+render_preview_list(selected_steps, steps_info)
 
 
 ########################################
-# OPERATIONS LIST MANAGEMENT
+# SELECTION LIST MANAGEMENT
 #
 # Add steps based on user selections.
 # User can add from checking in the steps table, but also there are some options to add dependencies, usages, etc.
 ########################################
 # Header
-render_operations_list(steps_df)
+render_selection_list(steps_df)
 
 
 ########################################
@@ -122,27 +128,4 @@ render_operations_list(steps_df)
 ########################################
 
 if st.session_state.selected_steps:
-    cols = st.columns(3, border=True)
-    ####################################################################################################################
-    # UPDATE STEPS
-    ####################################################################################################################
-    # Add an expander menu with additional parameters for the update command.
-    # with st.container(border=True):
-    with cols[0]:
-        render_action_update()
-
-    ####################################################################################################################
-    # EXECUTE SNAPSHOTS AND ETL STEPS
-    ####################################################################################################################
-    # Add an expander menu with additional parameters for the ETL command.
-    # with st.container(border=True):
-    with cols[1]:
-        render_action_execute(steps_df)
-
-    ####################################################################################################################
-    # ARCHIVE STEPS
-    ####################################################################################################################
-    # Add an expander menu with additional parameters for the ETL command.
-    # with st.container(border=True):
-    with cols[2]:
-        render_action_archive()
+    render_operations(steps_df)
