+++ apps/wizard/app_pages/dashboard/operations.py
@@ -1,173 +1,248 @@
+"""Optional actions to be performed on the selected steps."""
+
+from datetime import datetime
+
 import pandas as pd
 import streamlit as st
-from structlog import get_logger
 
-from apps.wizard.app_pages.dashboard.utils import NON_UPDATEABLE_IDENTIFIERS, _add_steps_to_operations, remove_step
+from apps.wizard.utils.cached import execute_bash_command
 from apps.wizard.utils.components import st_horizontal
+from etl.config import OWID_ENV
 
-log = get_logger()
-
-
-def render_operations_list(steps_df):
-    """Render operations list."""
-    with st_horizontal():
-        st.markdown("""### Operations list""")
-        st.markdown(f":primary-background[{len(st.session_state.selected_steps)} steps]")
-    with st.container(border=True):
-        # Create an operations list, that contains the steps (selected from the main steps table) we will operate upon.
-        # Note: Selected steps might contain steps other those selected in the main steps table, based on user selections (e.g. dependencies).
-        if st.session_state.selected_steps:
-            for index, step in enumerate(st.session_state.selected_steps):
-                # Define the layout of the list.
-                _show_row_with_step_details(steps_df, step, index)
-
-            # Show main buttons in the operations list
-            _show_main_operations_buttons(steps_df)
-
-        else:
-            st.markdown(":grey[_No rows selected for operation..._]")
+# Current date.
+# This is used as the default version of new steps to be created.
+TODAY = datetime.now().strftime("%Y-%m-%d")
 
 
-def _show_row_with_step_details(steps_df, step, index):
-    """Show row in the operations list with the details of the given step.
+def render_operations(steps_df):
+    st.markdown("### Operations")
+    cols = st.columns(3, border=True)
+    ####################################################################################################################
+    # UPDATE STEPS
+    ####################################################################################################################
+    # Add an expander menu with additional parameters for the update command.
+    # with st.container(border=True):
+    with cols[0]:
+        render_action_update()
 
-    # Define the columns in order (from left to right) as a list of tuples (message, key suffix, function).
+    ####################################################################################################################
+    # EXECUTE SNAPSHOTS AND ETL STEPS
+    ####################################################################################################################
+    # Add an expander menu with additional parameters for the ETL command.
+    # with st.container(border=True):
+    with cols[1]:
+        render_action_execute(steps_df)
 
-    TODO: Consider adding step buttons to:
-        * Execute ETL step for only the current step.
-        * Edit metadata for the current step.
-    """
-    col1, col2 = st.columns([2.5, 4])
-    with col1:
-        with st.container(height=40, border=False):
-            # with col1:
-            if step in st.session_state.selected_steps_table:
-                st.markdown(f"**{step.replace('data://', '')}**")
-            else:
-                st.markdown(f"{step.replace('data://', '')}")
-    with col2:
-        with st_horizontal(justify_content="space-between"):
-            actions = [
-                (
-                    "Add direct dependencies",
-                    "direct_dependencies",
-                    "Add direct dependencies of this step to the **Operations list**. Direct dependencies are steps that are loaded directly by the current step.",
-                ),
-                (
-                    "Add all dependencies",
-                    "all_active_dependencies",
-                    "Add all dependencies (including indirect dependencies) of this step to the **Operations list**. Indirect dependencies are steps that are needed, but not directly loaded, by the current step. In other words: dependencies of dependencies.",
-                ),
-                (
-                    "Add direct usages",
-                    "direct_usages",
-                    "Add direct usages of this step to the **Operations list**. Direct usages are those steps that load the current step directly.",
-                ),
-                (
-                    "Add all usages",
-                    "all_active_usages",
-                    "Add all usages (including indirect usages) of this step to the **Operations list**. Indirect usages are those steps that need, but do not directly load, the current step. In other words: usages of usages.",
-                ),
-            ]
-            unique_key = f"remove_{step}_{index}"
-            st.button(
-                label="🗑️",
-                key=unique_key,
-                on_click=lambda step=step: remove_step(step),
-                help="Remove this step from the **Operations list**.",
-                type="secondary",
+    ####################################################################################################################
+    # ARCHIVE STEPS
+    ####################################################################################################################
+    # Add an expander menu with additional parameters for the ETL command.
+    # with st.container(border=True):
+    with cols[2]:
+        render_action_archive()
+
+
+def render_action_update():
+    """Render container with "Update" action."""
+    st.markdown("##### Update steps")
+    with st_horizontal(vertical_alignment="center"):
+        with st.popover("More settings", icon=":material/settings:"):
+            dry_run_update = st.toggle(
+                "Dry run",
+                True,
+                help="If checked, the update command will not write anything to the DAG or create any files.",
             )
-            # Display the operations list.
-            for action_name, key_suffix, help_text in actions:
-                # Create a unique key for the button (if any button is to be created).
-                unique_key = f"{key_suffix}_{step}_{index}"
-                # Add related steps
-                st.button(
-                    label=f":blue[{action_name}]",
-                    key=unique_key,
-                    on_click=lambda step=step, key_suffix=key_suffix: _include_related_steps(
-                        steps_df, step, key_suffix
-                    ),
-                    help=help_text,
-                    type="tertiary",
-                )
-
-
-def _show_main_operations_buttons(steps_df):
-    """Show main buttons in the operations list:
-
-    - Clear list
-    - Remove non-updateable
-    - Replace with latest versions
-    """
-    with st_horizontal():
-        # Add button to clear the operations list.
-        st.button(
-            ":primary[:material/clear_all: Clear list]",
-            help="Remove all steps currently in the **Operations list**.",
-            # type="tertiary",
-            key="clear_operations_list",
-            on_click=lambda: st.session_state.selected_steps.clear(),
+            version_new = st.text_input("New version", value=TODAY, help="Version of the new steps to be created.")
+        btn_submit = st.button(
+            "Run",
+            help="Update all steps in the **Operations list**.",
+            type="primary",
+            icon=":material/play_circle:",
+            # use_container_width=True,
         )
 
-        st.button(
-            label=":primary[:material/delete_outline: Remove non-updateable (e.g. population)]",
-            help="Remove steps that cannot be updated (i.e. with `update_period_days=0`), and other auxiliary datasets, namely: "
-            + "\n- ".join(sorted(NON_UPDATEABLE_IDENTIFIERS)),
-            # type="tertiary",
-            key="remove_non_updateable",
-            on_click=lambda: _remove_non_updateable_steps(steps_df),
-        )
-
-        st.button(
-            label=":primary[:material/system_update_alt: Replace steps with their latest versions]",
-            help="Replace steps in the **Operations list** by their latest version available. You may want to use this button after updating steps, to be able to operate on the newly created steps.",
-            # type="tertiary",
-            key="replace_with_latest",
-            on_click=lambda: _upgrade_steps_in_operations_list(steps_df),
-        )
-
-
-def _include_related_steps(steps_df: pd.DataFrame, step: str, column_related: str):
-    """User can add additional steps to the operations list based on the selected step.
-
-    E.g. adding direct dependencies, all usages, etc.
-    """
-    steps_related = steps_df[steps_df["step"] == step]
-    if len(steps_related) == 0:
-        log.error(f"Step {step} not found in the steps table.")
-    elif len(steps_related) == 1:
-        steps_related = steps_df[steps_df["step"] == step][column_related].item()
-        # Add steps to operations list
-        _add_steps_to_operations(steps_related)
-    else:
-        st.error(f"More than one step found with the same URI {step}!")
-        st.stop()
-
-
-def _remove_non_updateable_steps(steps_df):
-    # Remove steps that cannot be updated (because update_period_days is set to 0).
-    # For convenience, also remove steps that a user most likely doesn't want to update.
-    non_updateable_steps = steps_df[
-        (steps_df["update_period_days"] == 0) | (steps_df["identifier"].isin(NON_UPDATEABLE_IDENTIFIERS))
-    ]["step"].tolist()
-    st.session_state.selected_steps = [
-        step for step in st.session_state.selected_steps if step not in non_updateable_steps
-    ]
-
-
-def _upgrade_steps_in_operations_list(steps_df):
-    new_list = []
-    for step in st.session_state.selected_steps:
-        step_info = steps_df[steps_df["step"] == step].iloc[0].to_dict()
-        step_identifier = step_info["identifier"]
-        latest_version = step_info["latest_version"]
-        step_latest = steps_df[(steps_df["identifier"] == step_identifier) & (steps_df["version"] == latest_version)][
-            "step"
-        ]
-        if not step_latest.empty:
-            new_list.append(step_latest.item())
+    # Button to execute the update command and show its output.
+    if btn_submit:
+        if OWID_ENV.env_local == "production":
+            st.error("The update command is not available in production. Update steps locally or in staging.")
+            st.stop()
         else:
-            new_list.append(step)
+            with st.spinner("Executing step updater..."):
+                # TODO: It would be better to directly use StepUpdater instead of a subprocess.
+                command = (
+                    "etl update "
+                    + " ".join(st.session_state.selected_steps)
+                    + " --non-interactive"
+                    + f" --step-version-new {version_new}"
+                )
+                if dry_run_update:
+                    command += " --dry-run"
+                cmd_output = execute_bash_command(command)
+                # Show the output of the command in an expander.
+                with st.expander("Command:", expanded=True):
+                    st.text(command)
+                    st.text_area("Output", value=cmd_output, height=300, key="cmd_output_area")
+                if "error" not in cmd_output.lower():
+                    # Celebrate that the update was successful, why not.
+                    st.balloons()
+                    if not dry_run_update:
+                        # Reload steps_df to include the new steps.
+                        st.session_state["reload_key"] += 1
+                # Add a button to close the output expander.
+                st.button("Close and reload _Steps table_", key="acknowledge_cmd_output")
 
-    st.session_state.selected_steps = new_list
+
+def render_action_execute(steps_df: pd.DataFrame):
+    st.markdown("##### Execute steps")
+    with st_horizontal(vertical_alignment="center"):
+        with st.popover("More settings", icon=":material/settings:"):
+            dry_run_etl = st.toggle(
+                "Dry run",
+                True,
+                help="If checked, no snapshots will be executed, and ETL will be executed in dry-run mode.",
+            )
+            force_only = st.toggle(
+                "Force run",
+                False,
+                help="If checked, the ETL steps will be forced to be executed (even if they are already executed).",
+            )
+            run_snapshots = st.toggle(
+                "Run snapshot scripts",
+                False,
+                help="If checked, run snapshot scripts (if any in the **Operations list**).",
+            )
+            run_grapher = st.toggle(
+                "Run grapher steps",
+                False,
+                help="If checked, run grapher steps with --grapher (if any in the **Operations list**).",
+            )
+
+        btn_etl_run = st.button(
+            label="Run",
+            help="Run ETL on all data steps in the **Operations list** (and optionally also execute snapshots).",
+            type="primary",
+            icon=":material/play_circle:",
+            # use_container_width=True,
+        )
+
+        # Button to execute the update command and show its output.
+        if btn_etl_run:
+            if OWID_ENV.env_local == "production":
+                st.error("Running the ETL is not available in production. Run them locally or in staging.")
+                st.stop()
+            else:
+                with st.spinner("Executing ETL..."):
+                    command = _define_command_to_execute_snapshots_and_etl_steps(
+                        steps_df=steps_df,
+                        dry_run=dry_run_etl,
+                        force_only=force_only,
+                        run_snapshots=run_snapshots,
+                        run_grapher=run_grapher,
+                    )
+                    cmd_output = execute_bash_command(cmd=command)
+                    # Show the output of the command in an expander.
+                    with st.expander("Command:", expanded=True):
+                        st.text(command)
+                        st.text_area("Output", value=cmd_output, height=300, key="cmd_output_area")
+                    if "error" not in cmd_output.lower():
+                        # Celebrate that the update was successful, why not.
+                        st.balloons()
+                    # Add a button to close the output expander.
+                    st.button("Close", key="acknowledge_cmd_output_etl_run")
+
+
+def render_action_archive():
+    st.markdown("##### Archive steps")
+    with st_horizontal(vertical_alignment="center"):
+        with st.popover("More settings", icon=":material/settings:"):
+            dry_run_archive = st.toggle(
+                "Dry run",
+                True,
+                help="If checked, nothing will be written to the dag.",
+            )
+            include_usages_archive = st.toggle(
+                "Include usages",
+                True,
+                help="If checked, archive also other archivable steps using the steps selected.",
+            )
+
+        btn_archive = st.button(
+            "Run",
+            help="Move archivable steps in the **Operations list** to their corresponding archive dag.",
+            type="primary",
+            icon=":material/play_circle:",
+            # use_container_width=True,
+        )
+
+    # Button to execute the update command and show its output.
+    if btn_archive:
+        if OWID_ENV.env_local == "production":
+            st.error("Archiving is not available in production. Run them locally or in staging.")
+            st.stop()
+        else:
+            with st.spinner("Archiving steps..."):
+                command = "etl archive " + " ".join(st.session_state.selected_steps) + " --non-interactive"
+                if dry_run_archive:
+                    command += " --dry-run"
+                if include_usages_archive:
+                    command += " --include-usages"
+                cmd_output = execute_bash_command(command)
+                # Show the output of the command in an expander.
+                with st.expander("Command:", expanded=True):
+                    st.text(command)
+                    st.text_area("Output", value=cmd_output, height=300, key="cmd_output_area")
+                if "error" not in cmd_output.lower():
+                    # Celebrate that the operation was successful.
+                    st.balloons()
+                    if not dry_run_archive:  # NOTE: this was previously `dry_run_update`. Replace this back to it if unexpected errors appear.
+                        # Reload steps_df to include the new steps.
+                        st.session_state["reload_key"] += 1
+                # Add a button to close the output expander.
+                st.button("Close and reload _Steps table_", key="acknowledge_cmd_output")
+
+
+def _define_command_to_execute_snapshots_and_etl_steps(
+    steps_df: pd.DataFrame,
+    dry_run: bool = True,
+    force_only: bool = False,
+    run_snapshots: bool = False,
+    run_grapher: bool = False,
+):
+    # Execute ETL for all steps in the operations list.
+    snapshot_steps = [step for step in st.session_state.selected_steps if step.startswith("snapshot://")]
+    etl_steps = [step for step in st.session_state.selected_steps if not step.startswith("snapshot://")]
+
+    command = ""
+    if run_snapshots:
+        # First write a command that will attempt to run all snapshots sequentially.
+        for snapshot_step in snapshot_steps:
+            # Identify script for current snapshot.
+            script = steps_df[steps_df["step"] == snapshot_step]["full_path_to_script"].item()
+            # Define command to be executed.
+            command += f"python {script} && "
+
+        if dry_run:
+            # If dry_run, we do not want to execute the command, but simply print it.
+            command = f"echo '{command}' && "
+
+    if etl_steps:
+        # Then let ETL run all remaining steps (ETL will decide the order).
+        # Define command to be executed.
+        command += f"etl run {' '.join(etl_steps)} "
+
+        if dry_run:
+            command += " --dry-run"
+
+        if force_only:
+            command += " --force --only"
+
+        if run_grapher:
+            # To run grapher steps (i.e. grapher://grapher/... steps) we need to remove the "data://" at the
+            # beginning of the step name, otherwise, grapher://grapher/... steps will be ignored.
+            command = command.replace("data://grapher/", "grapher/")
+            command += " --grapher"
+
+    if command.endswith("&& "):
+        command = command[:-3]
+
+    return command
